python/requirements.txt
Flask
Flask-SQLAlchemy
oss2
flask_migrate
pyjwt